Rename diff wrapper and move class to the diff package
